+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Windows ARM64 Editor support
+- Apple tvOS support
+- Apple visionOS support
 
 ### Changed
 - Updated Google Draco to version 1.5.7.